Add information on connecting to wifi networks with special character passwords
At least on my S5 (FW2034/2023.05.0), ";" is a valid password character. When adding a second S5 to my home, I discover that since I added the first one there has been a check added to Valetudo that ensures the wifi password does not have any special characters in it. As I knew that at least for my robot, in my setup, ";" is not an issue (as my other robot has been working fine for years) I found the miio command to set the credentials and all went smoothly from that point.